Refactor common code out of platform-specific implementation
